Remove darwina part from crab service
